fix(issue.service): make fine persistence even after returned #6
Previously, fine is calculated at runtime and send it to the client. The
problem with this approach is that, even after returning the book, the
fine is calculated at the runtime.

To solve this issue, the fine is persisted in the database after the
book is returned by the student.